Concurrency improvement

+++ JavaNeutrons/src/main/java/org/eastsideprep/javaneutrons/core/Part.java
@@ -313,8 +313,6 @@ public class Part {
     // exposure measurements in sieverts
     //
     double getSieverts(String type) {
-
-        
-        return this.fluenceMap.get(type).totalSieverts;
+        return this.fluenceMap.get(type).totalSieverts.doubleValue();
     }
 }
